requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
opencv-python>=4.6.0
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
pyyaml>=6.0
albumentations>=1.2.0
tensorboard>=2.8.0
wandb>=0.12.0
xmltodict>=0.12.0
lxml>=4.8.0
ultralytics>=8.0.0
gradio>=4.0.0
